# apps/armi-runtime/src/armi_runtime/interfaces/creator_web_resources/openapi.json
{
  "components": {
    "schemas": {
...
      "EffectResponse": {
        "additionalProperties": false,
        "properties": {
          "action_intent_ref": {
            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
            "title": "Action Intent Ref",
            "type": "string"
          },
          "action_intent_revision_ref": {
            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
            "title": "Action Intent Revision Ref",
            "type": "string"
          },
          "attempt_count": {
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Attempt Count",
            "type": "integer"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "pattern": "\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:[0-5]\\d\\.\\d{6}Z",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "capability_kind": {
            "enum": [
              "creator.scene.reply",
              "codex.delegated-work"
            ],
            "title": "Capability Kind",
            "type": "string"
          },
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "current_attempt_no": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Attempt No"
          },
          "current_attempt_ref": {
            "anyOf": [
              {
                "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Attempt Ref"
          },
          "current_dispatch_state": {
            "anyOf": [
              {
                "enum": [
                  "prepared",
                  "dispatching",
                  "settled"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Dispatch State"
          },
          "current_observation_ref": {
            "anyOf": [
              {
                "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Observation Ref"
          },
          "effect_id": {
            "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
            "title": "Effect Id",
            "type": "string"
          },
          "effect_kind": {
            "enum": [
              "creator_response",
              "codex_delegation"
            ],
            "title": "Effect Kind",
            "type": "string"
          },
          "last_observation_kind": {
            "anyOf": [
              {
                "enum": [
                  "receipt",
                  "query",
                  "rejection",
                  "ambiguous",
                  "runner_verified",
                  "runner_failed",
                  "runner_unknown",
                  "runner_cancelled"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Observation Kind"
          },
          "last_observation_reliability": {
            "anyOf": [
              {
                "enum": [
                  "reliable",
                  "operator_attested",
                  "inconclusive"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Observation Reliability"
          },
          "observation_conclusion": {
            "anyOf": [
              {
                "enum": [
                  "completed",
                  "failed",
                  "unknown",
                  "cancelled"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observation Conclusion"
          },
          "observation_evidence_kind": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observation Evidence Kind"
          },
          "observation_reason": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observation Reason"
          },
          "projection_version": {
            "const": "creator-effect.v6",
            "title": "Projection Version",
            "type": "string"
          },
          "registered_at": {
            "pattern": "\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:[0-5]\\d\\.\\d{6}Z",
            "title": "Registered At",
            "type": "string"
          },
          "response_text": {
            "anyOf": [
              {
                "maxLength": 65536,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Text"
          },
          "settled_at": {
            "anyOf": [
              {
                "pattern": "\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:[0-5]\\d\\.\\d{6}Z",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settled At"
          },
          "status": {
            "enum": [
              "registered",
              "dispatching",
              "completed",
              "failed",
              "unknown",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "verification_action": {
            "anyOf": [
              {
                "enum": [
                  "verify_local_inbox",
                  "verify_external_delivery",
                  "verify_codex_result"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Action"
          },
          "verification_status": {
            "enum": [
              "not_started",
              "pending",
              "verified",
              "operator_attested",
              "inconclusive"
            ],
            "title": "Verification Status",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "projection_version",
          "effect_id",
          "action_intent_ref",
          "action_intent_revision_ref",
          "capability_kind",
          "effect_kind",
          "status",
          "verification_status",
          "registered_at",
          "attempt_count"
        ],
        "title": "EffectResponse",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "RejectedOutcomeResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDescriptorResponse"
          },
          "message": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "occurred_at": {
            "pattern": "\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:[0-5]\\d\\.\\d{6}Z",
            "title": "Occurred At",
            "type": "string"
          },
          "status": {
            "const": "rejected",
            "title": "Status",
            "type": "string"
          },
          "trace_id": {
            "pattern": "[0-9a-f]{32}",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "trace_id",
          "occurred_at",
          "message",
          "status",
          "error"
        ],
        "title": "RejectedOutcomeResponse",
        "type": "object"
      },
...
      "UnavailableOutcomeResponse": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": "1.0",
            "title": "Contract Version",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDescriptorResponse"
          },
          "message": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "occurred_at": {
            "pattern": "\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:[0-5]\\d\\.\\d{6}Z",
            "title": "Occurred At",
            "type": "string"
          },
          "recovery_hint": {
            "anyOf": [
              {
                "maxLength": 4096,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recovery Hint"
          },
          "status": {
            "const": "unavailable",
            "title": "Status",
            "type": "string"
          },
          "trace_id": {
            "pattern": "[0-9a-f]{32}",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "trace_id",
          "occurred_at",
          "message",
          "status",
          "error"
        ],
        "title": "UnavailableOutcomeResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "browserSessionBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "ARMI Creator Interface",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/effects/{effect_id}": {
      "get": {
        "operationId": "getEffect",
        "parameters": [
          {
            "in": "path",
            "name": "effect_id",
            "required": true,
            "schema": {
              "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
              "title": "Effect Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EffectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnavailableOutcomeResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "browserSessionBearer": []
          }
        ],
        "summary": "Get Effect"
      }
    },
    "/v1/effects/{effect_id}/artifacts/{artifact_kind}": {
      "get": {
        "operationId": "getEffectArtifact",
        "parameters": [
          {
            "in": "path",
            "name": "effect_id",
            "required": true,
            "schema": {
              "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
              "title": "Effect Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_kind",
            "required": true,
            "schema": {
              "enum": [
                "patch",
                "final_result",
                "validation_report"
              ],
              "title": "Artifact Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Explicit verified Codex result artifact."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnavailableOutcomeResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "browserSessionBearer": []
          }
        ],
        "summary": "Get Effect Artifact"
      }
    },
...
    "/v1/operations/{result_ref}": {
      "get": {
        "operationId": "getCreatorOperation",
        "parameters": [
          {
            "in": "path",
            "name": "result_ref",
            "required": true,
            "schema": {
              "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}",
              "title": "Result Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationOutcomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnavailableOutcomeResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "browserSessionBearer": []
          }
        ],
        "summary": "Get Creator Operation"
      }
    },
...
    "/v1/scenes/{scene_key}/codex-tasks": {
      "post": {
        "operationId": "acceptCreatorCodexTask",
        "parameters": [
          {
            "in": "path",
            "name": "scene_key",
            "required": true,
            "schema": {
              "pattern": "[a-z0-9][a-z0-9._-]{0,63}",
              "title": "Scene Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "pattern": "[A-Za-z0-9][A-Za-z0-9._:-]{0,127}",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatorCodexTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedOutcomeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RejectedOutcomeResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnavailableOutcomeResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "browserSessionBearer": []
          }
        ],
        "summary": "Accept Creator Codex Task"
      }
    },
...
  }
}